+++ src/colors/amber.ts
