+++ src/colors/crimson.ts
