+++ src/colors/cyan.ts
