+++ src/colors/gold.ts
