+++ src/colors/gray.ts
