--- src/colors/index.js
@@ -1,73 +1,0 @@
-// grays
-import gray from './gray.js';
-import mauve from './mauve.js';
-import slate from './slate.js';
-import sage from './sage.js';
-import olive from './olive.js';
-import sand from './sand.js';
-
-//
-import tomato from './tomato.js';
-
-import red from './red.js';
-import ruby from './ruby.js';
-import crimson from './crimson.js';
-import pink from './pink.js';
-import plum from './plum.js';
-import purple from './purple.js';
-import violet from './violet.js';
-
-import iris from './iris.js';
-import indigo from './indigo.js';
-import blue from './blue.js';
-import cyan from './cyan.js';
-import teal from './teal.js';
-import jade from './jade.js';
-import green from './green.js';
-import grass from './grass.js';
-import brown from './brown.js';
-import bronze from './bronze.js';
-import gold from './gold.js';
-import sky from './sky.js';
-import mint from './mint.js';
-import lime from './lime.js';
-import yellow from './yellow.js';
-import amber from './amber.js';
-import orange from './orange.js';
-
-const colors = {
-    gray: gray,
-    mauve: mauve,
-    slate: slate,
-    sage: sage,
-    olive: olive,
-    sand: sand,
-    tomato: tomato,
-    red: red,
-    ruby: ruby,
-    crimson: crimson,
-    pink: pink,
-    plum: plum,
-    purple: purple,
-    violet: violet,
-    iris: iris,
-    indigo: indigo,
-    blue: blue,
-    cyan: cyan,
-    teal: teal,
-    jade: jade,
-    green: green,
-    grass: grass,
-    brown: brown,
-    bronze: bronze,
-    gold: gold,
-    sky: sky,
-    mint: mint,
-    lime: lime,
-    yellow: yellow,
-    amber: amber,
-    orange: orange,
-
-};
-
-export default colors;
