+++ src/colors/index.ts
@@ -1,0 +1,68 @@
+import gray from './gray';
+import mauve from './mauve';
+import slate from './slate';
+import sage from './sage';
+import olive from './olive';
+import sand from './sand';
+import tomato from './tomato';
+import red from './red';
+import ruby from './ruby';
+import crimson from './crimson' 
+import pink from './pink';
+import plum from './plum'
+import purple from './purple';
+import violet from './violet';
+import iris from './iris';
+import indigo from './indigo';
+import blue from './blue';
+import cyan from './cyan';
+import teal from './teal';
+import jade from './jade';
+import green from './green';
+import grass from './grass';
+import brown from './brown';
+import bronze from './bronze';
+import gold from './gold';
+import sky from './sky';
+import mint from './mint';
+import lime from './lime';
+import yellow from './yellow';
+import amber from './amber';
+import orange from './orange';
+
+const colors = {
+    gray: gray,
+    mauve: mauve,
+    slate: slate,
+    sage: sage,
+    olive: olive,
+    sand: sand,
+    tomato: tomato,
+    red: red,
+    ruby: ruby,
+    crimson: crimson,
+    pink: pink,
+    plum: plum,
+    purple: purple,
+    violet: violet,
+    iris: iris,
+    indigo: indigo,
+    blue: blue,
+    cyan: cyan,
+    teal: teal,
+    jade: jade,
+    green: green,
+    grass: grass,
+    brown: brown,
+    bronze: bronze,
+    gold: gold,
+    sky: sky,
+    mint: mint,
+    lime: lime,
+    yellow: yellow,
+    amber: amber,
+    orange: orange,
+
+} as const;
+
+export default colors;
