+++ src/colors/lime.ts
