+++ src/colors/mauve.ts
