+++ src/colors/pink.ts
