+++ src/colors/plum.ts
