+++ src/colors/ruby.ts
