+++ src/colors/sky.ts
