+++ src/colors/slate.ts
