+++ src/colors/teal.ts
