+++ src/colors/yellow.ts
